% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLD_scrape_wahlrechtde.R
\name{germanpolls}
\alias{germanpolls}
\title{A function to scrape Wahlrecht.de}
\usage{
germanpolls(region = "de")
}
\description{
germanpolls() is a function that gets polling data from [Wahlrecht.de](http://www.wahlrecht.de) for seven polling organisations and ten parties.
}
\examples{
polling_data <- germanpolls() 

}
