% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_from_xml_de.R
\name{get_data_from_xml_de}
\alias{get_data_from_xml_de}
\title{Get data of Germany}
\usage{
get_data_from_xml_de()
}
\value{
dataframe
}
\description{
Function to get data from German wide polls: http://www.wahlrecht.de/umfragen/bundesweite.xml
}
