% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_from_xml_laender.R
\name{get_data_from_xml_laender}
\alias{get_data_from_xml_laender}
\title{Get data of Bundesländer}
\usage{
get_data_from_xml_laender(url)
}
\arguments{
\item{url}{url of XML file, auto-generated from `region`}
}
\value{
data frame
}
\description{
Function to get data from states.
}
