% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germanpolls.R
\name{germanpolls}
\alias{germanpolls}
\title{Get data of German polls from wahlrecht.de.}
\usage{
germanpolls(region = "de")
}
\arguments{
\item{region}{string, all available regions are documented in the readme}
}
\value{
dataframe
}
\description{
This is the main function to parse XML files from Wahlrecht.de to extract polling data. 
It catches all XML elements that can be found.
Not all regions have XML available, their data can't be accessed with this function.
}
\examples{
germanpolls(region = "de")
germanpolls(region = "by")
}
